Clean up model creation with model parameters
See how we bind the specials to the params?  Those *same* params will be
bound we the make query functions are called on demand.  That is how it
should be (of course) and was a big mess when everything was just a
special variable (public or otherwise).

The neat trick is the INSERT call at the botton of create-model.  We are
eating our own dog food by keeping model parameters in our own model
table!