model parameters jdoc and jdoc-type are too much
We can't (easily) write raw select statements without hardcoding
'jdoc so we might as well remove any notion that it can be varied:
The column containing the JSON data in all our model relations is called
'jdoc and we're done.

So to, it has type jsonb (rather than json) because you can't do much
with the json type.