Add some CLOS types for models
This is the first in a long series of commits migrating Postgres-JSON
from "model as symbol" to "model as first class type".

The commits are broken up for logical comprehension.  Do not expect any
particular HEAD in this series of commits to actually build, it may or
may not until all commits in the 'CLOS redev' series are applied.